requirements/base.txt
# Django & DRF
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.2
django-cors-headers==4.3.1

# Environment variables
python-decouple==3.8
python-dotenv==1.0.0


# API documentation
drf-spectacular==0.29.0

# Database
# mysqlclient==2.2.1
psycopg2-binary==2.9.11 # Alternative: PostgreSQL